+++ test/learn/fakeClassificationLearningEnvironment.h
@@ -54,7 +54,7 @@ class FakeClassificationLearningEnvironment
   public:
     FakeClassificationLearningEnvironment()
         : ClassificationLearningEnvironment(3), data(1), value{0} {};
-    void doAction(uint64_t actionId) override
+    void doAction(double actionId) override
     {
         // Increment classificationTable
         ClassificationLearningEnvironment::doAction(actionId);
