+++ test/learn/stickGameAdversarial.cpp
@@ -48,7 +48,7 @@ Learn::LearningEnvironment* StickGameAdversarial::clone() const
     return (Learn::LearningEnvironment*)new StickGameAdversarial(*this);
 }
 
-void StickGameAdversarial::doAction(uint64_t actionID)
+void StickGameAdversarial::doAction(double actionID)
 {
     LearningEnvironment::doAction(actionID);
