+++ test/learn/stickGameAdversarial.h
@@ -96,7 +96,7 @@ class StickGameAdversarial : public Learn::AdversarialLearningEnvironment
     virtual LearningEnvironment* clone() const override;
 
     // Inherited via LearningEnvironment
-    virtual void doAction(uint64_t actionID) override;
+    virtual void doAction(double actionID) override;
 
     // Inherited via LearningEnvironment
     virtual void reset(size_t seed = 0,
